Blub Blub Blub
Idk I kinda implemented some parts of it.
Some notes are:
the allocation reuse in the builder is pointless, as there are no real
values that are discarded when finalized, i.e. all allocations will not
be freed, but rather moved.
I think I have now implemented the brunt of the requirements of the
format, so all that is left now is to implement the instructions
and their construction, then it's on to the codegen part.