FUCK
For one I did some refactoring and realized JaggedVec is pointless and
we can just use a KeyVec with 2 indices stored.
For two: the problem is that liveness can change conditionally between
blocks. This is problematic as it means we can not store it as some
use-attribute. While this is fine as we *could* instead store it as a
searchable array of `(Instruction, Liveness)` that is separated into
chunks for each block, the problem is then figuring out said liveness.
Liveness can have 3 states: either a given instruction is dead in a
given block (which can be implicitly assumed if no other state is
present), it can be alive for the entire duration of the block (this
is if *any* of the following blocks use the value, even if they are
only conditionally branched to) or it can be aive for part of the
block, in which case if dies at some (last-use) instruction.
The problem with determining liveness is that it means first
annotating it at block granularity (dead/alive, can be annotated
during the instruction insertiong) before moving onto instruction
granularity (partial), which causes problems however, as it means each
block's chunk fragments.